Moved get_class_name() to path, so we can reuse it in database package.
