+++ src/astro/databases/__init__.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from airflow.hooks.base import BaseHook
 
 from astro.databases.base import BaseDatabase
-from astro.utils.path import get_dict_with_module_names_to_dot_notations
+from astro.utils.path import get_class_name, get_dict_with_module_names_to_dot_notations
 
 DEFAULT_CONN_TYPE_TO_MODULE_PATH = get_dict_with_module_names_to_dot_notations(
     Path(__file__)
@@ -29,7 +29,6 @@ def create_database(conn_id: str) -> BaseDatabase:
     conn_type = BaseHook.get_connection(conn_id).conn_type
     module_path = CONN_TYPE_TO_MODULE_PATH[conn_type]
     module = importlib.import_module(module_path)
-    module_name = module_path.split(".")[-1]
-    class_name = module_name.title() + "Database"
+    class_name = get_class_name(module_ref=module, suffix="Database")
     database: BaseDatabase = getattr(module, class_name)(conn_id)
     return database
