+++ src/astro/files/locations/__init__.py
@@ -1,10 +1,10 @@
 import importlib
 from pathlib import Path
-from typing import Any, Optional
+from typing import Optional
 
 from astro.constants import FileLocation
 from astro.files.locations.base import BaseFileLocation
-from astro.utils.path import get_dict_with_module_names_to_dot_notations
+from astro.utils.path import get_class_name, get_dict_with_module_names_to_dot_notations
 
 DEFAULT_CONN_TYPE_TO_MODULE_PATH = get_dict_with_module_names_to_dot_notations(
     Path(__file__)
@@ -25,29 +25,3 @@ def create_file_location(path: str, conn_id: Optional[str] = None) -> BaseFileLo
     class_name = get_class_name(module_ref)
     location: BaseFileLocation = getattr(module_ref, class_name)(path, conn_id)
     return location
-
-
-def get_class_name(module_ref: Any, suffix: str = "Location") -> str:
-    """Get class name to be dynamically imported. Class name are expected to be in following formats
-    example -
-    module name: test
-    suffix: Abc
-
-    expected class names -
-        1. TESTAbc
-        2. TestAbc
-    :param module_ref: Module from which to get class location type implementation
-    """
-
-    module_name = module_ref.__name__.split(".")[-1]
-    class_names_formats = [
-        f"{module_name.title()}{suffix}",
-        f"{module_name.upper()}{suffix}",
-    ]
-    for class_names_format in class_names_formats:
-        if hasattr(module_ref, class_names_format):
-            return class_names_format
-
-    raise ValueError(
-        "No expected class name found, please note that the class names should an expected formats."
-    )
