+++ src/astro/utils/path.py
@@ -1,5 +1,5 @@
 import pathlib
-from typing import Dict
+from typing import Any, Dict
 
 
 def get_module_dot_notation(module_path: pathlib.Path) -> str:
@@ -50,3 +50,30 @@ def get_dict_with_module_names_to_dot_notations(
                 module_path
             )
     return module_name_to_dot_notation
+
+
+def get_class_name(module_ref: Any, suffix: str = "Location") -> str:
+    """Get class name to be dynamically imported. Class name are expected to be in following formats
+    example -
+    module name: test
+    suffix: Abc
+
+    expected class names -
+        1. TESTAbc
+        2. TestAbc
+    :param module_ref: Module from which to get class location type implementation
+    :param suffix: suffix for class name
+    """
+
+    module_name = module_ref.__name__.split(".")[-1]
+    class_names_formats = [
+        f"{module_name.title()}{suffix}",
+        f"{module_name.upper()}{suffix}",
+    ]
+    for class_names_format in class_names_formats:
+        if hasattr(module_ref, class_names_format):
+            return class_names_format
+
+    raise ValueError(
+        "No expected class name found, please note that the class names should an expected formats."
+    )
